requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.2
kiwisolver==1.4.7
matplotlib==3.9.3
MouseInfo==0.1.3
numpy==2.2.0
packaging==24.2
pillow==11.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyobjc-core==10.3.2
pyobjc-framework-ApplicationServices==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreText==10.3.2
pyobjc-framework-Quartz==10.3.2
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
rubicon-objc==0.4.9
six==1.17.0
